initial commit
Comments are very useful in any language to tell user what is the task of any function or operation. The comments are neglected by the compiler, so whatever you write in the comments won't have any effect on the working of a code. In Java, comments can be written as mentioned below:
/*This is a comment*/
//This is also a comment
You are given a complete function that outputs a, b, and c. Comment the code that outputs b, so only a and c gets printed.

Example 1: 

Input:
a = 5, b = 6, c = 15
Output:
5
15

Example 2: 

Input:
a = 5, b = 8, c = 15
Output:
5
15

 

User Task:
Your task is to comment the unnecessary statement.

 

Constraints:
1 <= a <= 106
1 <= b <= 106
1 <= c<= 106