Frequency Count | Java
Given an ArrayList arr of n elements, you need to find frequency of each element.

Example 1 :

Input:
6
10 5 20 5 10 5

Output:
2 3 1

Explanation: 
The freq are as follows: 10:2, 5:3, 20:1, so we print 
the freq in the order of appearance of the elements. 
We print freq of any element only once.

Example 2 :

Input:
4
100 100 100 100

Output:
4

Explanation: 
Here 100 appears 4 times.

Your Task:
You need to complete the function frequencyCount that takes an ArrayList and its size as arguments. The function returns an ArrayList that contains the frequencies of the elements.

Constraints:
1 <= n <= 106
0 <= arri <= 106