List of Smaller Elements | Java
You are given an ArrayList arr of n items. You are also given an integer k. You need to find all the elements in arr that are smaller than k.

Example 1:

Input:
n = 6
arr = {8, 100, 20, 40, 3, 7}
k = 10
Output:
8 3 7
Explanation:
Here k is 10. From left-to-right the
elements less than 10 are 8 3 7.

Example 2:

Input:
n = 6
arr = {100, 20, 40, 60, 80, 200}
k = 50
Output:
20 40
Explanation:
Here k is 50. From left-to-right the
elements less than 50 are 20 40.

Your Task:
Since this is a function problem, you don't need to take any input. Just complete the function getSmaller that takes arr and k as arguments and returns a list of elements in arr that are smaller than k. The returned list should contain the elements in the left-to-right order. The printing is done by the driver code.

Constraints:
1 <= n <= 107
0 <= k, arri <= 107